requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-leaflet==0.28.3
GDAL==3.6.2
idna==3.4
numpy @ file:///D:/bld/numpy_1675642697443/work
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyproj==3.4.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
